% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMGnify.R
\name{addMGnify}
\alias{addMGnify}
\alias{addMGnify,SummarizedExperiment,MultiAssayExperiment-method}
\alias{addMGnify,SummarizedExperiment,SummarizedExperiment-method}
\alias{addMGnify,SummarizedExperiment,ANY-method}
\title{Add results from MGnifyR to HoloFoodR results}
\usage{
addMGnify(x, y, ...)

\S4method{addMGnify}{SummarizedExperiment,MultiAssayExperiment}(
  x,
  y,
  exp.name1 = "metagenomic",
  exp.name2 = "metagenomic_amplicon",
  replace = TRUE,
  ...
)

\S4method{addMGnify}{SummarizedExperiment,SummarizedExperiment}(x, y, ...)

\S4method{addMGnify}{SummarizedExperiment,ANY}(x, y, id.col1 = "sample_biosample", id.col2 = "accession", ...)
}
\arguments{
\item{x}{\code{SummarizedExperiment}. Results from
\code{MGnifyR::getResult()}.}

\item{y}{\code{MultiAssayExperiment} or \code{SummarizedExperiment} or
\code{data.frame}-like table. Results from \code{HoloFoodR::getResult()} or
sample metadata from it.}

\item{...}{optional arguments not used currently.}

\item{exp.name1}{\code{Character scalar}. Specifies the name of experiment
that will be added to \code{y}. (Default: \code{"metagenomic"})}

\item{exp.name2}{\code{Character scalar}. Specifies the name of experiment
from HoloFoodR results. This experiment is used to match IDs with MGnify
data. (Default: \code{"metagenomic_amplicon"})}

\item{replace}{\code{Logical scalar}. Whether to replace the template
experiment. (Default: \code{TRUE})}

\item{id.col1}{\code{Character scalar}. Specifies the name of column from
\code{colData(x)} that includes HoloFood identifiers.
(Default: \code{"sample_biosample"})}

\item{id.col2}{\code{Character scalar}. Specifies the name of column
from \code{colData(y[[exp.name2]])} that includes HoloFood identifiers.
(Default: \code{"accession"})}
}
\value{
\code{MultiAssayExperiment}
}
\description{
Add results from MGnifyR to HoloFoodR results
}
\details{
Metagenomic data is located in MGnify instead of HoloFoodR. However, the
databases use different sample identifiers. With this function, you can
merge MGnify results to your HoloFoodR results. The function combined sample
metadata as some data is located in HoloFood and some in MGnify.
}
\examples{

# mae <- addMGnify(tse, mae)

}
