% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoloFoodR.R
\docType{package}
\name{HoloFoodR-package}
\alias{HoloFoodR}
\alias{HoloFoodR-package}
\title{\code{HoloFoodR} package}
\description{
\code{HoloFoodR} implements an interface to the EBI HoloFood database.
See the vignette for a general introduction to this package,
[about HoloFood](https://www.holofood.eu/) for general HoloFood
information, and
[API documentation](https://www.holofooddata.org/api/docs) for
details on the JSONAPI implementation.
}
\seealso{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssaySummarizedExperiment}
}
\author{
\strong{Maintainer}: Tuomas Borman \email{tuomas.v.borman@utu.fi} (\href{https://orcid.org/0000-0002-8563-8884}{ORCID})

Authors:
\itemize{
  \item Leo Lahti \email{leo.lahti@iki.fi} (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
}

}
