% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoloFoodR.R
\docType{_PACKAGE}
\name{HoloFoodR}
\alias{HoloFoodR}
\title{\code{HoloFoodR} package}
\description{
\code{HoloFoodR} implements an interface to the EBI HoloFood database.
See the vignette for a general introduction to this package,
[about HoloFood](https://www.holofood.eu/) for general HoloFood
information, and
[API documentation](https://www.holofooddata.org/api/docs) for
details on the JSONAPI implementation.
}
\seealso{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssaySummarizedExperiment}
}
