% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaboLights.R
\name{getMetaboLights}
\alias{getMetaboLights}
\title{Get metabolomic data from MetaboLights database}
\usage{
getMetaboLights(url, ...)
}
\arguments{
\item{url}{\code{character vector} specifying the URL address of study in
MetaboLights database.}

\item{...}{optional arguments:
\itemize{
  
  \item \strong{cache.dir} \code{Character scalar} specifying directory
  where downloaded file is stored. (Default: \code{tempdir()})
  
}}
}
\value{
\code{list}
}
\description{
Get metabolomic data from MetaboLights database
}
\details{
The HoloFood database primarily comprises targeted metabolomic data,
omitting non-targeted metabolomic information. Nonetheless, it features URLs
linking to studies within the MetaboLights database. This functionality
enables users to access non-targeted metabolomic data. The function returns
a structured list encompassing data frames for study metadata,
assay metadata, and assay.
}
\examples{

\dontrun{
url <- "https://www.ebi.ac.uk/metabolights/ws/studies/MTBLS4381"
res <- getMetaboLights(url)
names(res)
head(res[["feat_meta"]])
}

}
\seealso{
\code{\link[HoloFoodR:getResult]{getResult}}
\code{\link[HoloFoodR:getData]{getData}}
}
