% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaboLights.R
\name{getMetaboLights}
\alias{getMetaboLights}
\alias{getMetaboLightsFile}
\title{Get metabolomic data from MetaboLights database}
\usage{
getMetaboLights(study.id, ...)

getMetaboLightsFile(study.id, file, ...)
}
\arguments{
\item{study.id}{\code{character vector} specifying the study identifier of
data that is going to be fetched from the MetaboLights database.}

\item{...}{optional arguments:
\itemize{
  
  \item \strong{cache.dir} \code{Character scalar} specifying directory
  where downloaded file is stored. (Default: \code{tempdir()})
  
  \item \strong{timeout} \code{Integer scalar} specifying timeout
   in seconds for loading the file. (Default: \code{5*60})
  
}}

\item{file}{\code{character vector} specifying the files that are being
fetched.}
}
\value{
\code{list}
}
\description{
Get metabolomic data from MetaboLights database
}
\details{
The HoloFood database primarily comprises targeted metabolomic data,
omitting non-targeted metabolomic information. Nonetheless, it features URLs
linking to studies within the MetaboLights database. This functionality
enables users to access non-targeted metabolomic data. The
\code{getMetaboLights} function returns
a structured list encompassing processed data in \code{data.frame} format
for study metadata,
assay metadata, and assay. The metadata includes the file names of raw
spectra data. Those files can be loaded with \code{getMetaboLightsFile}.
}
\examples{

# This example is not run, because the server fails to respond sometimes.
if( FALSE ){
    res <- getMetaboLights("MTBLS4381")
    file_paths <- getMetaLightsFile(
        study.id = "MTBLS4381",
        file = res[["assay_meta"]][["Raw Spectral Data File"]]
        )
}

}
\seealso{
\code{\link[HoloFoodR:getResult]{getResult}}
\code{\link[HoloFoodR:getData]{getData}}
}
