% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaboLights.R
\name{getMetaboLights}
\alias{getMetaboLights}
\title{Get metabolomic data from MetaboLights database}
\usage{
getMetaboLights(url, ...)
}
\arguments{
\item{url}{\code{character vector} specifying the URL address of study in
MetaboLights database.}

\item{...}{optional arguments:
\itemize{
  
  \item \strong{cache.dir} \code{Character scalar} specifying cache
  directory. (Default: \code{tempdir()})
  
}}
}
\value{
\code{list}
}
\description{
Get metabolomic data from MetaboLights database
}
\details{
Non-targeted metabolomics data is in MetaboLigths...
}
\seealso{
\code{\link[HoloFoodR:getResult]{getResult}}
\code{\link[HoloFoodR:getData]{getData}}
}
