% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoloFoodR.R
\docType{package}
\name{HoloFoodR}
\alias{HoloFoodR-package}
\alias{HoloFoodR}
\title{\code{HoloFoodR} package}
\description{
\code{HoloFoodR} implements an interface to the EBI HoloFood database.
See the vignette for a general introduction to this package,
[about HoloFood](https://www.holofood.eu/) for general HoloFood
information, and
[API documentation](https://docs.holofooddata.org/api.html) for
details on the JSONAPI implementation.
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment}
}
\author{
\strong{Maintainer}: Tuomas Borman \email{tuomas.v.borman@utu.fi} (\href{https://orcid.org/0000-0002-8563-8884}{ORCID})

Authors:
\itemize{
  \item Artur Sannikov \email{arsann@utu.fi} (\href{https://orcid.org/0000-0001-7765-123X}{ORCID})
  \item Leo Lahti \email{leo.lahti@iki.fi} (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
}

}
